% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derfinderData-package.R
\docType{package}
\name{derfinderData-package}
\alias{derfinderData}
\alias{derfinderData-package}
\title{derfinderData: Processed BigWigs from BrainSpan for examples}
\description{
Processed 22 samples from BrainSpan keeping only the information for chromosome 21. Data is stored in the BigWig format and is used for examples in other packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leekgroup/derfinderData}
  \item Report bugs at \url{https://github.com/leekgroup/derfinderData/issues}
}

}
\author{
\strong{Maintainer}: Leonardo Collado-Torres \email{lcolladotor@gmail.com} (\href{https://orcid.org/0000-0003-2140-308X}{ORCID})

Authors:
\itemize{
  \item Andrew Jaffe \email{andrew.jaffe@libd.org} (\href{https://orcid.org/0000-0001-6886-1454}{ORCID})
  \item Jeffrey Leek \email{jtleek@gmail.com} (\href{https://orcid.org/0000-0002-2873-2671}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
