% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derfinderData-package.R
\docType{package}
\name{derfinderHelper-package}
\alias{derfinderHelper-package}
\title{Processed BigWigs from BrainSpan for examples}
\description{
Processed 22 samples from BrainSpan keeping only the information for
chromosome 21. Data is stored in the BigWig format and is used for examples
in other packages.
}
\author{
Leonardo Collado-Torres \href{mailto:lcolladotor@gmail.com}{lcolladotor@gmail.com}
}
\keyword{package}
