\docType{package}
\name{derfinderData-package}
\alias{derfinderData-package}
\title{Simulated data for illustrating derfinder's functions}
\description{
Reads were generated for chromosome 22 hg19 transcripts
with the goal of illustrating how to use derfinder. The
design is a two group comparison with 10 samples per group
with 20% differentially expressed transcripts (10% higher
in each group). The package provides the fasta files as
well as the resulting BAM files from aligning the reads
with TopHat v2.0.9.
}
\author{
Leonardo Collado-Torres <lcollado@jhsph.edu>
}
\references{
Leonardo Collado-Torres, Alyssa Frazee, Andrew Jaffe and
Jeffrey Leek (2013). derfinder: Fast differential
expression analysis of RNA-seq data at base-pair
resolution. https://github.com/lcolladotor/derfinder
}
\keyword{package}

