% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{derfinderHelper-package}
\alias{derfinderHelper-package}
\title{Processed BigWigs from BrainSpan for examples}
\description{
Processed 22 samples from BrainSpan keeping only the information for
chromosome 21. Data is stored in the BigWig format and is used for examples #' in other packages.
}
\author{
Leonardo Collado-Torres <lcollado@jhu.edu>
}
\keyword{package}

